Launch puredata at boot
Systemd is configured to run a shell script that launches our patch.
The effect is that at boot, the input is routed to the output.  It is
also running through a low pass filter, but when the patch launches the
corner frequency is set high enough that it may as well be bypassed.

Power on until sound, approx 25 seconds.  Perhaps turbo start?